Struct ctf::parse::CtfParser

pub struct CtfParser<'data, Endian> { /* private fields */ }

Top-level parser.

Implementations

impl<'data> CtfParser<'data, ()>

pub fn new(
bytes: &'data [u8]
) -> Res<Either<BeCtfParser<'data>, LeCtfParser<'data>>>

Constructor.

Yields either a big-endian or a low-endian parser, based on the magic-number starting the sequence of bytes. This function is not meant to be used directly, use the parse macro instead, which hides the details of handling the Either part.

impl<'data, Endian> CtfParser<'data, Endian>

pub fn header(&self) -> &Ctf

Header accessor.

pub fn trace_info(&self) -> &Info<'data>

Trace info accessor.

impl<'data, Endian> CtfParser<'data, Endian> where
Parser<'data, Endian>: CanParse<'data>,

Pseudo-parsers: parses a very tiny amout of data to produce a subparser.

pub fn next_packet<'me>(
&'me mut self
) -> Res<Option<PacketParser<'me, 'data, Endian>>>

Yields a PacketParser for the next packet, if any.