Module ctf::prelude::uid

UID types for charts, filters and subfilters.

All UID types implement serialize and deserialize.

Types Chart, Filter and SubFilter are the straightforward UIDs. This module also has a Line type which augments Filter with two additional variants:

the “catch-all filter”, which is the filter that catches everything the other filters do not catch;

the “everything filter”, which is the filter that catches all allocations, independently of the user-defined filters.

Structs

Alloc

Allocation UID.

AllocMap

Map from allocation UIDs to something.

Chart

Chart UID.

Filter

Filter UID.

SubFilter

Sub-filter UID.

Enums

Line

A UID for a line in the chart.